GPCRAnalysis

Documentation for GPCRAnalysis.

GPCRAnalysis.align

GPCRAnalysis.chargelocations

GPCRAnalysis.columnwise_entropy

GPCRAnalysis.download_alphafolds

GPCRAnalysis.filter_long!

GPCRAnalysis.filter_species!

GPCRAnalysis.getchain

GPCRAnalysis.project_sequences

GPCRAnalysis.sortperm_msa

GPCRAnalysis.species

GPCRAnalysis.try_download_alphafold

GPCRAnalysis.uniprotX

GPCRAnalysis.align — Method

align(fixedpos::AbstractMatrix{Float64}, moving::AbstractVector{PDBResidue}, sm::SequenceMapping)
align(fixed::AbstractVector{PDBResidue}, moving::AbstractVector{PDBResidue}, sm::SequenceMapping)

Return a rotated and shifted version of moving so that the α-carbons of residues moving[sm] have least mean square error deviation from positions fixedpos or those of residues fixed.

GPCRAnalysis.chargelocations — Method

chargelocations(pdb::AbstractVector{PDBResidue}; include_his::Bool=false)

Return a list of potential charge locations in the protein structure. Each is a tuple (position, residueindex, AAname). The positions are those of N (in positively-charged residues like Arg & Lys) or O (in negatively-charged residues like Asp and Glu). N- and C-termini are not included in the list. While each residue will carry a net total charge of ±1, the location of each potential charge will be listed (1 for Lys, 2 each for Arg, Asp, and Glu).

By default, histidine is not considered charged, but you can include it by setting include_His=true.

GPCRAnalysis.columnwise_entropy — Function

columnwise_entropy(msa::AbstractMultipleSequenceAlignment, aacode = ReducedAlphabet("(AILMV)(NQST)(RHK)(DE)(FWY)CGP")

Compute the entropy of each column in an MSA. Low entropy indicates high conservation.

Unmatched entries ('-' residues) contribute to the entropy calculation as if they were an ordinary residue.

GPCRAnalysis.download_alphafolds — Method

download_alphafolds(msa::AbstractMultipleSequenceAlignment; dirname=pwd())

Download all available AlphaFold structures for the sequences in msa. Missing entries are silently skipped.

GPCRAnalysis.filter_long! — Method

filter_long!(msa::AbstractMultipleSequenceAlignment, minres::Real)

Remove all sequences from msa with fewer than minres matching residues.

GPCRAnalysis.filter_species! — Method

filter_species!(msa::AbstractMultipleSequenceAlignment, speciesname::AbstractString)

Remove all sequences from msa except those with species(sequencename) equal to speciesname.

GPCRAnalysis.getchain — Method

getchain(filename::AbstractString; model="1", chain="A")

Read a PDB file filename and extract the specified chain.

GPCRAnalysis.project_sequences — Method

X = project_sequences(msa::AbstractMultipleSequenceAlignment; fracvar::Real = 0.9)

Perform a classical multidimensional scaling analysis to project the sequences in msa to a space in which pairwise distances approximately reproduce 100 - percentsimilarity(seq1, seq2). The dimensionality is chosen to reconstruction fracvar of the variance.

GPCRAnalysis.sortperm_msa — Method

tour = sortperm_msa(msa::AbstractMultipleSequenceAlignment)

Order the sequences in msa to minimize the "tour length" visiting each sequence once. The length between sequences is defined at 100 - percentsimilarity(seq1, seq2).

This can be useful for graphical or alignment display by grouping obviously-similar sequences near one another.

GPCRAnalysis.species — Method

species(name)

Extract the species identifier from a UniProt "X_Y" entry or elaborated variant (e.g., PFAM sequence name).

See also uniprotX.

Examples

julia> species("Q8VGW6_MOUSE/31-308")
"MOUSE"

GPCRAnalysis.try_download_alphafold — Function

try_download_alphafold(uniprotXname, path=alphafoldfile(uniprotXname))

Attempt to download an AlphaFold structure. Returns nothing if no entry corresponding to uniprotXname exists; otherwise it return path, the pathname of the saved file.

GPCRAnalysis.uniprotX — Method

uniprotX(name)

Extract the UniProt "X" entry from an X_Y entry or elaborated variant (e.g., PFAM sequence name).

See also species.

Examples

julia> uniprotX("Q8VGW6_MOUSE/31-308")
"Q8VGW6"