GPCRAnalysis

Documentation for GPCRAnalysis.

GPCRAnalysis.BWScheme

GPCRAnalysis.StructAlign

GPCRAnalysis.align

GPCRAnalysis.chargelocations

GPCRAnalysis.chimerax_script

GPCRAnalysis.columnwise_entropy

GPCRAnalysis.download_alphafolds

GPCRAnalysis.filter_long!

GPCRAnalysis.filter_species!

GPCRAnalysis.getchain

GPCRAnalysis.lookupbw

GPCRAnalysis.project_sequences

GPCRAnalysis.residue_centroid

GPCRAnalysis.residue_centroid_matrix

GPCRAnalysis.residueindex

GPCRAnalysis.residueindex

GPCRAnalysis.sortperm_msa

GPCRAnalysis.species

GPCRAnalysis.try_download_alphafold

GPCRAnalysis.uniprotX

GPCRAnalysis.BWScheme — Method

BWScheme(conserved_idx, tmspans)

Specify the Ballesteros-Weinstein scheme used for a particular protein. conserved_idx is a list of 7 "most conserved" residues per helix (rhodopsin family: N1, D2, R3, W4, P5, P6, P7) and the span of each helix.

Examples

For mouse rhodopsin (P15409),

julia> opsd_scheme = BWScheme([55, 83, 135, 161, 215, 267, 303],
           [34:64, 73:99, 107:139, 150:173, 200:229, 246:277, 285:309]);

GPCRAnalysis.StructAlign — Method

StructAlign(struct1::AbstractVector{PDBResidue}, struct2::AbstractVector{PDBResidue}, filename::AbstractString)

Create a structure-based alignment between struct1 and struct2. filename is the name of the TM-align "results" file (e.g., https://zhanggroup.org//TM-align/example/873772.html).

See also residueindex.

GPCRAnalysis.align — Method

align(fixedpos::AbstractMatrix{Float64}, moving::AbstractVector{PDBResidue}, sm::SequenceMapping)
align(fixed::AbstractVector{PDBResidue}, moving::AbstractVector{PDBResidue}, sm::SequenceMapping)

Return a rotated and shifted version of moving so that the α-carbons of residues moving[sm] have least mean square error deviation from positions fixedpos or those of residues fixed.

GPCRAnalysis.chargelocations — Method

chargelocations(pdb::AbstractVector{PDBResidue}; include_his::Bool=false)

Return a list of potential charge locations in the protein structure. Each is a tuple (position, residueindex, AAname). The positions are those of N (in positively-charged residues like Arg & Lys) or O (in negatively-charged residues like Asp and Glu). N- and C-termini are not included in the list. While each residue will carry a net total charge of ±1, the location of each potential charge will be listed (1 for Lys, 2 each for Arg, Asp, and Glu).

By default, histidine is not considered charged, but you can include it by setting include_His=true.

GPCRAnalysis.chimerax_script — Method

chimerax_script(filename, uprot_list, msa::AnnotatedMultipleSequenceAlignment, colidxs;
                dir=pwd(), chain_transparency=80, styles=Dict{Int,String}(), extras=String[])

Create a chimerax visualization script with name filename. uprot_list is a list of UniProtX names that you want to visualize. The AlphaFold PDB files for these proteins should be dir. msa is a Multiple Sequence alignment and colidxs specifies the column indices in msa that you'd like to visualize.

chain_transparency sets the transparency on the ribbon diagrams (0 = not transparent). styles can be used to affect the display, e.g., Dict(k => "@SD sphere") would cause methionines at column index k to be displayed with the sulfur in sphere mode. extras can be used to hand-specify a number of additional commands; this can be useful if, for example, the msa has occasional misalignments.

Examples

Suppose you have the msa for rhodopsin (mouse: P15409), then:

chimerax_script("myscript.cxc", ["P15409"], msa, [i1, i2, i3])

where i1 through i3 are column-indices in the msa that you'd like to view.

GPCRAnalysis.columnwise_entropy — Function

columnwise_entropy(msa::AbstractMultipleSequenceAlignment, aacode = ReducedAlphabet("(AILMV)(NQST)(RHK)(DE)(FWY)CGP")

Compute the entropy of each column in an MSA. Low entropy indicates high conservation.

Unmatched entries ('-' residues) contribute to the entropy calculation as if they were an ordinary residue.

GPCRAnalysis.download_alphafolds — Method

download_alphafolds(msa::AbstractMultipleSequenceAlignment; dirname=pwd())

Download all available AlphaFold structures for the sequences in msa. Missing entries are silently skipped.

GPCRAnalysis.filter_long! — Method

filter_long!(msa::AbstractMultipleSequenceAlignment, minres::Real)

Remove all sequences from msa with fewer than minres matching residues.

GPCRAnalysis.filter_species! — Method

filter_species!(msa::AbstractMultipleSequenceAlignment, speciesname::AbstractString)

Remove all sequences from msa except those with species(sequencename) equal to speciesname.

GPCRAnalysis.getchain — Method

getchain(filename::AbstractString; model="1", chain="A")

Read a PDB file filename and extract the specified chain.

GPCRAnalysis.lookupbw — Method

helix, residue_position = lookupbw(idx::Integer, scheme::BWScheme)

Calculate the Ballesteros-Weinstein residue number coresponding to residue idx. tmspans describes the transmembrane regions in the reference, and bwconserved the index of the most-conserved residue.

Examples

For mouse rhodopsin (P15409),

julia> opsd_scheme = BWScheme([55, 83, 135, 161, 215, 267, 303],
           [34:64, 73:99, 107:139, 150:173, 200:229, 246:277, 285:309]);

julia> lookupbw(160, opsd_scheme)
(4, 49)

julia> lookupbw((4, 49), opsd_scheme)
160

This is the residue just before the most-conserved residue of helix 4.

GPCRAnalysis.project_sequences — Method

X = project_sequences(msa::AbstractMultipleSequenceAlignment; fracvar::Real = 0.9)

Perform a classical multidimensional scaling analysis to project the sequences in msa to a space in which pairwise distances approximately reproduce 100 - percentsimilarity(seq1, seq2). The dimensionality is chosen to reconstruction fracvar of the variance.

GPCRAnalysis.residue_centroid — Method

residue_centroid(r::PDBResidue)

Compute the mean position of all atoms in r excluding hydrogen.

Residue centers may yield a more reliable measure of "comparable residues" than the α-carbons (see CAmatrix) because they incorporate the orientation of the side chain relative to the overall fold.

See also residue_centroid_matrix.

GPCRAnalysis.residue_centroid_matrix — Method

residue_centroid_matrix(seq)

Return a matrix of all residue centroids as columns. See also residue_centroid.

GPCRAnalysis.residueindex — Method

residueindex(sa::StructAlign, idx1, nothing)
residueindex(sa::StructAlign, idx1, nothing, ±1)

Calculate the residue index in structure 2 corresponding to residue idx1 in structure 1. The second form allows you to find the nearest corresponding residue in the forward (+1) or reverse (-1) directions, if idx1 is not among the mapped residues.

GPCRAnalysis.residueindex — Method

residueindex(sa::StructAlign, nothing, idx2)
residueindex(sa::StructAlign, nothing, idx2, ±1)

Calculate the residue index in structure 1 corresponding to residue idx2 in structure 2. The second form allows you to find the nearest corresponding residue in the forward (+1) or reverse (-1) directions, if idx2 is not among the mapped residues.

GPCRAnalysis.sortperm_msa — Method

tour = sortperm_msa(msa::AbstractMultipleSequenceAlignment)

Order the sequences in msa to minimize the "tour length" visiting each sequence once. The length between sequences is defined at 100 - percentsimilarity(seq1, seq2).

This can be useful for graphical or alignment display by grouping obviously-similar sequences near one another.

GPCRAnalysis.species — Method

species(name)

Extract the species identifier from a UniProt "X_Y" entry or elaborated variant (e.g., PFAM sequence name).

See also uniprotX.

Examples

julia> species("Q8VGW6_MOUSE/31-308")
"MOUSE"

GPCRAnalysis.try_download_alphafold — Function

try_download_alphafold(uniprotXname, path=alphafoldfile(uniprotXname))

Attempt to download an AlphaFold structure. Returns nothing if no entry corresponding to uniprotXname exists; otherwise it return path, the pathname of the saved file.

GPCRAnalysis.uniprotX — Method

uniprotX(name)

Extract the UniProt "X" entry from an X_Y entry or elaborated variant (e.g., PFAM sequence name).

See also species.

Examples

julia> uniprotX("Q8VGW6_MOUSE/31-308")
"Q8VGW6"